feat(parse): add parse pkg
