+++ internal/parse/testdata/openapi3.yml
